+++ Changelog.rst
@@ -4,6 +4,11 @@ ApexEvents
 
 Changelog
 -----------
+``0.4.1``
+    | Feature: [LEVEL9] Adds a clean UI to the output of the complete leaderboard
+    | Update: [LEVEL9] Adds a 'tournament concluded' chat message
+    | Update: [GENERAL] Code cleanup
+
 ``0.4.0``
     | Feature: [SUMMIT] Adds basic support for 'THE SUMMIT' event mode
     | Fix: [LEVEL9] Typo in function call ``Command.add_params``
@@ -12,7 +17,7 @@ Changelog
 ``0.3.5``
     | Update: [LEVEL9] Modifies ``/lvl9rank`` command for getting the entire leaderboard via an
       optional argument
-    | Update: Code cleanup
+    | Update: [GENERAL] Code cleanup
     | Fix: [LEVEL9] Plugin closes a LEVEL9 event one map earlier than expected
 
 ``0.3.0``